tutoring_agent.py
from flask import Flask, request, jsonify

tutoring_app = Flask(__name__)

@tutoring_app.route("/tutor", methods=["POST"])
def tutor():
    data = request.get_json()
    task = data.get("task")
    parameters = data.get("parameters", {})
    # Implement tutoring logic here, e.g., using OpenAI API
    response_content = f"Tutoring agent received task: {task} with params {parameters}"
    return jsonify({"status": "success", "response": response_content})

if __name__ == "__main__":
    tutoring_app.run(host="0.0.0.0", port=5001, debug=True)