requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
astroid==3.3.8
astunparse==1.6.3
bcrypt==4.2.1
bidict==0.23.1Flask
Flask-JWT-Extended
Flask-Cors
Flask-SQLAlchemy
itsdangerous
requests
openai
Werkzeug
psycopg2-binary